Feat: coloca autenticação nos métodos para 'role'

+++ src/services/roleService.ts
@@ -1,5 +1,6 @@
 import { Roles } from "../entities/roles";
 import { AppDataSource } from "../db_connection";
+import { ITokenRequest, TokenService } from "./tokenService";
 
 interface IRolesRequestName{
     role_name: string
@@ -16,7 +17,18 @@ interface IRolesRequestFull{
 
 export class RoleService{
     repository = AppDataSource.getRepository(Roles)
-    async novaRole({role_name}:IRolesRequestName){
+    async novaRole({role_name}:IRolesRequestName, {token}:ITokenRequest){
+        // verifica o token
+        const user = await new TokenService().verificaToken({token})
+        if (!user){ 
+            throw new Error(String("Você não tem autorização"))
+        }
+
+        // ERRO: não tem permissão (só usuários com role==1 podem alterar uma role)
+        if(user.fk_role_id.role_id!=1){
+            throw new Error("Você não tem permissão")
+        }
+
         // ERRO: o campo é obrigatório
         if(!role_name){
             throw new Error("'role_name' é obrigatório")
@@ -41,7 +53,18 @@ export class RoleService{
         return role
     }
 
-    async atualizaRole({role_id, role_name}:IRolesRequestFull){
+    async atualizaRole({role_id, role_name}:IRolesRequestFull, {token}:ITokenRequest){
+        // verifica o token
+        const user = await new TokenService().verificaToken({token})
+        if (!user){ 
+            throw new Error(String("Você não tem autorização"))
+        }
+
+        // ERRO: não tem permissão (só usuários com role==1 podem alterar uma role)
+        if(user.fk_role_id.role_id!=1){
+            throw new Error("Você não tem permissão")
+        }
+
         // ERRO: todos os campos são obrigatórios
         if(!role_id){
             throw new Error("'role_id' é obrigatório")
@@ -74,7 +97,18 @@ export class RoleService{
         return roleAtualizada
     }
 
-    async pegaUmaRolePeloId({role_id}:IRolesRequestId){
+    async pegaUmaRolePeloId({role_id}:IRolesRequestId, {token}:ITokenRequest){
+        // verifica o token
+        const user = await new TokenService().verificaToken({token})
+        if (!user){ 
+            throw new Error(String("Você não tem autorização"))
+        }
+
+        // ERRO: não tem permissão (só usuários com role==1 podem alterar uma role)
+        if(user.fk_role_id.role_id!=1){
+            throw new Error("Você não tem permissão")
+        }
+
         // SUCESSO
         const busca = await this.repository.findOne({where: {role_id: role_id}})
         if(!busca){
@@ -84,7 +118,18 @@ export class RoleService{
         }
     }
 
-    async pegaUmaRolePeloNome({role_name}:IRolesRequestName){
+    async pegaUmaRolePeloNome({role_name}:IRolesRequestName, {token}:ITokenRequest){
+        // verifica o token
+        const user = await new TokenService().verificaToken({token})
+        if (!user){ 
+            throw new Error(String("Você não tem autorização"))
+        }
+
+        // ERRO: não tem permissão (só usuários com role==1 podem alterar uma role)
+        if(user.fk_role_id.role_id!=1){
+            throw new Error("Você não tem permissão")
+        }
+
         // SUCESSO
         const busca = await this.repository.findOne({where: {role_name: role_name}})
         if(!busca){
@@ -94,9 +139,20 @@ export class RoleService{
         }
     }
 
-    async apagaUmaRolePeloId({role_id}:IRolesRequestId){
+    async apagaUmaRolePeloId({role_id}:IRolesRequestId, {token}:ITokenRequest){
+        // verifica o token
+        const user = await new TokenService().verificaToken({token})
+        if (!user){ 
+            throw new Error(String("Você não tem autorização"))
+        }
+
+        // ERRO: não tem permissão (só usuários com role==1 podem alterar uma role)
+        if(user.fk_role_id.role_id!=1){
+            throw new Error("Você não tem permissão")
+        }
+
         // SUCESSO
-        const busca = await this.pegaUmaRolePeloId({role_id})
+        const busca = await this.pegaUmaRolePeloId({role_id}, {token})
         if(busca){
             await this.repository.delete({role_id})
                 .catch(() => {
@@ -108,9 +164,20 @@ export class RoleService{
         }
     }
 
-    async apagaUmaRolePeloName({role_name}:IRolesRequestName){
+    async apagaUmaRolePeloName({role_name}:IRolesRequestName, {token}:ITokenRequest){
+        // verifica o token
+        const user = await new TokenService().verificaToken({token})
+        if (!user){ 
+            throw new Error(String("Você não tem autorização"))
+        }
+
+        // ERRO: não tem permissão (só usuários com role==1 podem alterar uma role)
+        if(user.fk_role_id.role_id!=1){
+            throw new Error("Você não tem permissão")
+        }
+
         // SUCESSO
-        const busca = await this.pegaUmaRolePeloNome({role_name})
+        const busca = await this.pegaUmaRolePeloNome({role_name}, {token})
         if(busca){
             await this.repository.delete({role_name: role_name})
                 .catch(() => {
